model/requirements.txt
fastapi
uvicorn
torch
torchvision
open-clip-torch
pymongo
pillow
nltk
scikit-learn
python-dotenv
numpy 
opencv-python
redis
celery
flower